config/dockerexec: Store container in config
Previously, when associating a session with a docker container the
container was found twice. This could technically result in a race
condition when a container was restarting right when the user was dis-
and reconnecting.

This commit updates the behaviour to instead store the associated
container in the context of the connection.